move az helper functions to arizoner cleaner
