+++ utils/arizona.py
@@ -1,13 +1,7 @@
 import pandas as pd
-
+import uuid
 from utils.clean import StateCleaner
 from utils.cleaner_utils import (
-    az_employment_checker,
-    az_individuals_convert,
-    az_name_clean,
-    az_organizations_convert,
-    az_transactions_convert,
-    az_transactor_sorter,
     convert_date,
     transactions_splitter,
 )
@@ -15,12 +9,337 @@ from utils.constants import (
     AZ_INDIVIDUALS_FILEPATH,
     AZ_ORGANIZATIONS_FILEPATH,
     AZ_TRANSACTIONS_FILEPATH,
+    state_abbreviations,
 )
+
+
+def az_name_clean(df: pd.DataFrame) -> pd.DataFrame:
+    """Replace empty 'candidate' value with 'committee_name' value
+
+    Because in some cases the 'candidate' column of the tables,
+    which, contrary to expected naming convention, contains the name of
+    the entity, is sometimes blank while the entity's name is in the
+    commitee_name column, we replace the relevant empty values
+    by their imputed values
+
+    args: df: detailed information dataframe
+
+    returns: modified detailed information dataframe
+
+    """
+
+    df_working = df.copy()
+
+    df_working["candidate"] = df.apply(
+        lambda row: row["committee_name"]
+        if (row["candidate"] == ("" or None or "" or """"""))
+        else row["candidate"],
+        axis=1,
+    )
+
+    return df_working
+
+
+def az_transactions_convert(df: pd.DataFrame) -> pd.DataFrame:
+    """Make raw transactions table schema-compliant
+
+    We take the relevant columns of the raw transactions
+    table and extract, reorder and relabel them
+    in compliance with the transactions table database schema
+
+    args: df: raw ransactions dataframe
+
+    returns: schema-compliant transactions dataframe
+    """
+
+    d = {
+        "transaction_id": df["PublicTransactionId"].astype(int).astype(str),
+        "donor_id": df["base_transactor_id"].astype(int).astype(str),
+        "year": df["TransactionDateYear"].astype(int),
+        "amount": df["Amount"].abs().astype(float),
+        "recipient_id": df["other_transactor_id"].astype(int).astype(int),
+        "office_sought": df["office_sought"].astype(str).str.lower(),
+        "purpose": df["Memo"].astype(str).str.lower(),
+        "transaction_type": df["TransactionType"].astype(str).str.lower(),
+        "TransactionTypeDispositionId": df["TransactionTypeDispositionId"],
+    }
+
+    trans_df = pd.DataFrame(data=d)
+
+    trans_df = trans_df.apply(az_donor_recipient_director, axis=1)
+
+    trans_df = trans_df.drop(columns=["TransactionTypeDispositionId"])
+
+    return trans_df
+
+
+def az_individuals_convert(details_df: pd.DataFrame) -> pd.DataFrame:
+    """Make individuals detail table schema compliant
+
+    INCOMPLETE
+
+    We take the relevant columns of the raw individual details
+    table and extract, reorder and relabel them in
+    compliance with the individuals table database schema
+
+    Presently, the function does not attempt to split names,
+    and instead leaves those columns blank and preserves the
+    available full names
+
+    NOTE: Employers information is presently difficult to
+    extract owing to inconsistent id locations across
+    various kinds of tables. A fix is underway.
+
+    args: df: transactions dataframe and individual details dataframe
+
+    returns: schema-compliant individual details dataframe
+
+    """
+
+    details_df = details_df.apply(az_individual_name_checker, axis=1)
+    details_df["full_name"] = details_df["full_name"].str.replace("\t", "")
+
+    employer = details_df["company"]
+
+    entity_type = details_df["entity_type"]
+
+    states_list = []
+
+    for i in details_df["committee_address"].str.split(" "):
+        if i is not None:
+            try:
+                abb = i[-2]
+                if " " + abb.upper() + " " not in state_abbreviations:
+                    states_list.append(None)
+                else:
+                    states_list.append(i[-2].upper())
+            except TypeError:
+                states_list.append(None)
+        else:
+            states_list.append(None)
+
+    d = {
+        "id": details_df["retrieved_id"].astype(int).astype(str),
+        "first_name": None,
+        "last_name": None,
+        "full_name": details_df["full_name"].astype(str).str.lower(),
+        "entity_type": entity_type.astype(str).str.lower(),
+        "state": states_list,
+        "party": details_df["party_name"].astype(str).str.lower(),
+        "company": employer.astype(str).str.lower(),
+    }
+
+    return pd.DataFrame(data=d)
+
+
+def az_organizations_convert(df: pd.DataFrame) -> pd.DataFrame:
+    """Make organizations detail table schema compliant
+
+    INCOMPLETE
+
+    We take the relevant columns of the raw organizations details
+    table and extract, reorder and relabel them
+    in compliance with the organizations table database schema
+
+    args: df: raw organizations dataframe
+
+    returns: schema-compliant organizations dataframe
+
+    """
+
+    entity_type = df["entity_type"]
+
+    states_list = []
+
+    for i in df["committee_address"].str.split(" "):
+        if i is not None:
+            try:
+                abb = i[-2]
+                if " " + abb.upper() + " " not in state_abbreviations:
+                    states_list.append(None)
+                else:
+                    states_list.append(i[-2].upper())
+            except TypeError:
+                states_list.append(None)
+        else:
+            states_list.append(None)
+
+    d = {
+        "id": df["retrieved_id"].astype(int).astype(str),
+        "name": df["committee_name"].astype(str).str.lower(),
+        "state": states_list,
+        "entity_type": entity_type.astype(str).str.lower(),
+    }
+
+    return pd.DataFrame(data=d)
+
+
+def az_transactor_sorter(row):
+    """Sorts ids into base and other transactor
+
+    Because the arizona transactions dataset records the ids
+    of transactors differently depending on the entity type,
+    it is necessary to sort through it and clarify the ids.
+    The base transactor is the entity in the transaction from
+    whom the transaction was collected, and the other transactor
+    is the other entity involved in the transaction
+
+    args: row of the transactions dataframe
+
+    returns: adjusted row of the transactions dataframe,
+    with new columns 'base_transactor_id' and
+    'other_transactor_id'
+
+    """
+    if row["entity_type"] in ["Vendor", "Individual"]:
+        row["base_transactor_id"] = row["TransactionNameGroupId"]
+        row["other_transactor_id"] = row["CommitteeId"]
+    else:
+        row["base_transactor_id"] = row["CommitteeId"]
+        row["other_transactor_id"] = row["TransactionNameGroupId"]
+
+    return row
+
+
+def az_donor_recipient_director(row):
+    """Sorts ids into donor and recipient columns
+
+    We switch the elements of a row in the transactions table if the
+    'TransactionTypeDispositionId' column indicates that the transaction
+    involves the other transactor receiving money and the base transactor
+    giving money. This is coded as a '2' in the dataset.
+
+    args: row of the transactions dataframe
+
+    returns: adjusted row of the transactions dataframe
+
+    """
+    if row["TransactionTypeDispositionId"] == 2:
+        row["donor_id"], row["recipient_id"] = row["recipient_id"], row["donor_id"]
+    return row
+
+
+def az_employment_checker(row, transactions):
+    """Retrieves employment data
+
+    We attempt to collect employment data for the
+    individuals dataframe. Because of how candidates vs
+    individual contributors are coded, it is necessary to
+    code the candidate employment separately, otherwise
+    the column shapes become different
+
+    args: row of the individuals dataframe,
+    transactions dataframe
+
+    returns: adjusted row of the individuals dataframe
+
+    """
+    if row["entity_type"] == "Candidate":
+        row["company"] = "None (Is a Candidate)"
+    else:
+        row["company"] = transactions[
+            transactions["retrieved_id"] == row["retrieved_id"]
+        ].iloc[0]["TransactionEmployer"]
+
+    return row
+
+
+def az_individual_name_checker(row):
+    """Collect names for individuals
+
+    Since names are coded differently for candidates vs
+    individual contributors, we go row by row and collect the
+    right element in each case
+
+    args: row of the individuals dataframe
+
+    returns: adjusted row of the individual dataframe
+    with the correct names of individual contributors and
+    candidates in the 'full name' column
+
+    """
+    if row["entity_type"] == "Candidate":
+        row["full_name"] = row["candidate"]
+    else:
+        row["full_name"] = row["retrieved_name"]
+    return row
+
+
+def az_id_table(
+    individuals_df, organizations_df, transactions_df, *args: int
+) -> pd.DataFrame:
+    """Create a table of old ids and new uuids"""
+
+    ind_ids = individuals_df["id"]
+
+    org_ids = organizations_df["id"]
+
+    trans_ids = transactions_df["transaction_id"]
+
+    years = pd.concat[
+        individuals_df["year"], organizations_df["year"], transactions_df["year"]
+    ]
+
+    all_ids = pd.concat([ind_ids, org_ids, trans_ids])
+
+    id_types = (
+        ["Individual"] * len(ind_ids)
+        + ["Organization"] * len(org_ids)
+        + ["Transaction"] * len(trans_ids)
+    )
+
+    d = {
+        "state": "AZ",
+        "year": years,
+        "entity_type": id_types,
+        "provided_id": all_ids,
+    }
+
+    id_map_df = pd.DataFrame(data=d)
+    id_map_df["database_id"] = [uuid.uuid4() for _ in range(len(id_map_df.index))]
+
+    return id_map_df
 
 
 class ArizonaCleaner(StateCleaner):
     """This class is based on the StateCleaner abstract class,
     and cleans Arizona data"""
+
+    def clean_state(self) -> (pd.DataFrame, pd.DataFrame, pd.DataFrame):
+        """Calls the other methods in order
+
+        This is the master function of the ArizonaCleaner
+        class, and calling it will activate the cleaning
+        pipeline which takes in filenames and outputs cleaned,
+        standardized, and schema-compliant tables
+
+        args: list of three filepaths which lead to dataframes
+
+        returns: three schema-compliant tables for
+        transactions, individuals, and organizations
+
+        """
+
+        filepaths = self.get_filepaths()
+
+        individuals, organizations, transactions = self.preprocess(filepaths)
+        transactions = transactions.head(1000)
+
+        details = pd.concat([individuals, organizations])
+
+        cleaned_transactions, cleaned_details = self.clean([transactions, details])
+
+        standardized_transactions, standardized_details = self.standardize(
+            [cleaned_transactions, cleaned_details]
+        )
+
+        (
+            az_individuals,
+            az_organizations,
+            az_transactions,
+        ) = self.create_tables([standardized_transactions, standardized_details])
+
+        return (az_individuals, az_organizations, az_transactions)
 
     def get_filepaths(self):
         return [
@@ -52,41 +371,6 @@ class ArizonaCleaner(StateCleaner):
             df_list.append(pd.read_csv(filepath))
 
         return df_list
-
-    def clean_state(self) -> (pd.DataFrame, pd.DataFrame, pd.DataFrame):
-        """Calls the other methods in order
-
-        This is the master function of the ArizonaCleaner
-        class, and calling it will activate the cleaning
-        pipeline which takes in filenames and outputs cleaned,
-        standardized, and schema-compliant tables
-
-        args: list of three filepaths which lead to dataframes
-
-        returns: three schema-compliant tables for
-        transactions, individuals, and organizations
-
-        """
-
-        filepaths = self.get_filepaths()
-
-        individuals, organizations, transactions = self.preprocess(filepaths)
-
-        details = pd.concat([individuals, organizations])
-
-        cleaned_transactions, cleaned_details = self.clean([transactions, details])
-
-        standardized_transactions, standardized_details = self.standardize(
-            [cleaned_transactions, cleaned_details]
-        )
-
-        (
-            az_individuals,
-            az_organizations,
-            az_transactions,
-        ) = self.create_tables([standardized_transactions, standardized_details])
-
-        return (az_individuals, az_organizations, az_transactions)
 
     def create_tables(
         self,
@@ -148,7 +432,6 @@ class ArizonaCleaner(StateCleaner):
         returns: details dataframe with relevant entity type
         names replaced by those for the regular schema
         """
-
         transactions_df, details_df = details_df_list[0], details_df_list[1]
 
         az_entity_name_dictionary = {
@@ -175,32 +458,26 @@ class ArizonaCleaner(StateCleaner):
         args: list of transactions and details dataframes
 
         returns: cleaned transactions and details dataframes
-
         """
-
-        transactions, details = data
-
-        merged_df = pd.merge(details, transactions, on="retrieved_id", how="inner")
-
-        # Filter rows in the first dataframe based on the common 'ids'
-        details = details[details["retrieved_id"].isin(merged_df["retrieved_id"])]
+        transactions, entities = data
 
         try:
             transactions["TransactionDate"] = transactions["TransactionDate"].apply(
                 convert_date
             )
-        except TypeError:
+        except TypeError as e:
+            print(f"Error converting Arizona dates: {e}")
             transactions["TransactionDate"] = transactions["TransactionDate"]
 
-        details = az_name_clean(details)
+        entities = az_name_clean(entities)
 
-        details = details.apply(az_employment_checker, args=(transactions,), axis=1)
+        entities = entities.apply(az_employment_checker, args=(transactions,), axis=1)
 
         transactions = transactions.apply(az_transactor_sorter, axis=1)
 
         merged_df = pd.merge(
             transactions["base_transactor_id"],
-            details[["retrieved_id", "office_name"]],
+            entities[["retrieved_id", "office_name"]],
             how="left",
             left_on="base_transactor_id",
             right_on="retrieved_id",
@@ -210,4 +487,4 @@ class ArizonaCleaner(StateCleaner):
 
         transactions["office_sought"] = office_sought
 
-        return [transactions, details]
+        return [transactions, entities]
